tdf: support longer time periods
Support time periods of up to 68 minutes in the time array by reserving
the top bit of the period as a scale factor. The reduces the maximum
value that can be represented at full resolution from 0.999 seconds to
0.499 seconds, but increases the maximum representable period from 0.999
seconds to 68 minutes. The resolution at the higher time scale is
125 ms.

+++ include/infuse/data_logger/high_level/tdf.h
@@ -66,7 +66,7 @@ enum {
  * @retval -errno Error code from @a tdf_add or @a tdf_data_logger_flush on error
  */
 int tdf_data_logger_log_array_dev(const struct device *dev, uint16_t tdf_id, uint8_t tdf_len,
-				  uint8_t tdf_num, uint64_t time, uint16_t period,
+				  uint8_t tdf_num, uint64_t time, uint32_t period,
 				  const void *data);
 
 /**
@@ -81,7 +81,7 @@ int tdf_data_logger_log_array_dev(const struct device *dev, uint16_t tdf_id, uin
  * @param data TDF data array
  */
 void tdf_data_logger_log_array(uint8_t logger_mask, uint16_t tdf_id, uint8_t tdf_len,
-			       uint8_t tdf_num, uint64_t time, uint16_t period, const void *data);
+			       uint8_t tdf_num, uint64_t time, uint32_t period, const void *data);
 
 /**
  * @brief Add a single TDF to a data logger
